refactor: split player zone subcomponents
